# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: e-IRF Electronic Incident Records Form API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error or wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /incidents/{id}/evidence:
    get:
      tags: [incidents]
      operationId: listEvidenceFiles
      summary: List evidence files for an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvidenceFile'
    post:
      tags: [incidents]
      operationId: addEvidenceFile
      summary: Record an uploaded evidence file for an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceFileInput'
      responses:
        "201":
          description: Evidence file recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceFile'
  /incidents/{id}/evidence/{fileId}:
    delete:
      tags: [incidents]
      operationId: deleteEvidenceFile
      summary: Remove an evidence file record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence file removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [officer]
      properties:
        officer:
          $ref: "#/components/schemas/Officer"
    OfficerRole:
      type: string
      enum: [admin, officer]
...
    Officer:
      type: object
      required: [id, name, badgeNumber, rank, role, username, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        badgeNumber:
          type: string
        rank:
          type: string
        role:
          $ref: "#/components/schemas/OfficerRole"
        username:
          type: string
        createdAt:
          type: string
...
    ChangePasswordBody:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    EvidenceFile:
      type: object
      required: [id, incidentId, fileName, objectPath, contentType, uploadedById, createdAt]
      properties:
        id:
          type: integer
        incidentId:
          type: integer
        fileName:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        fileSize:
          type: ["integer", "null"]
        sha256:
          type: ["string", "null"]
          pattern: '^[a-f0-9]{64}$'
          description: SHA-256 fingerprint calculated from the stored file bytes
        uploadedById:
          type: ["integer", "null"]
        uploadedByName:
          type: ["string", "null"]
        createdAt:
          type: string
    EvidenceFileInput:
      type: object
      required: [fileName, objectPath, contentType]
      properties:
        fileName:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        fileSize:
          type: ["integer", "null"]
...